# sstable: sstable.Writer optimizations