* sstable: encode varints directly into buf in blockWriter.store
* sstable: micro-optimize Writer.addPoint()
* sstable: minor cleanup of Writer/blockWriter
* sstable: micro-optimize blockWriter.store

Combined, the optimizations result in a decent improvement on the new
`sstable.Writer` benchmark.

```
name       old time/op  new time/op  delta
Writer-16  94.9ms ± 1%  74.6ms ± 4%  -21.40%  (p=0.000 n=10+20)
```